+++ xic/include/grip.h
@@ -179,17 +179,17 @@ private:
 
 struct cdelt_t
 {
-    unsigned long tab_key()  const { return (cd_key); }
+    uintptr_t tab_key()            const { return (cd_key); }
     cdelt_t *tab_next()            { return (cd_next); }
     cdelt_t *tgen_next(bool)       { return (cd_next); }
     void set_tab_next(cdelt_t *n)  { cd_next = n; }
 
-    void set_key(const void *p)    { cd_key = (unsigned long)p; }
+    void set_key(const void *p)    { cd_key = (uintptr_t)p; }
     itable_t<idelt_t> *table()     { return (cd_tab); }
     void set_table(itable_t<idelt_t> *t) { cd_tab = t; }
 
 private:
-    unsigned long cd_key;
+    uintptr_t cd_key;
     cdelt_t *cd_next;
     itable_t<idelt_t> *cd_tab;
 };
